[Frontend][last/5] Improve pooling entrypoints | clean up. (#39675)
