+++ vllm/entrypoints/pooling/embed/protocol.py
@@ -18,14 +18,15 @@ from pydantic import BaseModel, Field
 from vllm import PoolingParams
 from vllm.config import ModelConfig
 from vllm.entrypoints.openai.engine.protocol import OpenAIBaseModel, UsageInfo
-from vllm.entrypoints.pooling.base.protocol import (
+from vllm.renderers import TokenizeParams
+from vllm.utils import random_uuid
+
+from ..base.protocol import (
     ChatRequestMixin,
     CompletionRequestMixin,
     EmbedRequestMixin,
     PoolingBasicRequestMixin,
 )
-from vllm.renderers import TokenizeParams
-from vllm.utils import random_uuid
 
 # ---------------------------------------------------------------------------
 # OpenAI /v1/embeddings — request models
