+++ vllm/entrypoints/pooling/pooling/io_processor.py
@@ -4,12 +4,12 @@ from collections.abc import Sequence
 from typing import Any
 
 from vllm import PoolingParams, PoolingRequestOutput
-from vllm.entrypoints.pooling.base.io_processor import PoolingIOProcessor
 from vllm.inputs import EngineInput
 from vllm.logger import init_logger
 from vllm.plugins.io_processors import get_io_processor
 from vllm.renderers.inputs.preprocess import parse_model_prompt, prompt_to_seq
 
+from ..base.io_processor import PoolingIOProcessor
 from ..typing import OfflineInputsContext, OfflineOutputsContext, PoolingServeContext
 from .protocol import IOProcessorRequest, IOProcessorResponse
 
@@ -17,6 +17,8 @@ logger = init_logger(__name__)
 
 
 class PluginWithoutIOProcessorPlugins(PoolingIOProcessor):
+    # Some models, such as Terratorch (tests/models/test_terratorch.py),
+    # use plugin tasks in the pooler but do not use IO Processor plugins.
     name = "plugin"
