+++ vllm/entrypoints/pooling/typing.py
@@ -9,29 +9,30 @@ from fastapi import Request
 from pydantic import ConfigDict
 
 from vllm import PoolingParams, PoolingRequestOutput, PromptType
-from vllm.entrypoints.pooling.classify.protocol import (
+from vllm.inputs import DataPrompt, EngineInput
+from vllm.lora.request import LoRARequest
+
+from .classify.protocol import (
     ClassificationChatRequest,
     ClassificationCompletionRequest,
     ClassificationResponse,
 )
-from vllm.entrypoints.pooling.embed.protocol import (
+from .embed.protocol import (
     CohereEmbedRequest,
     EmbeddingBytesResponse,
     EmbeddingChatRequest,
     EmbeddingCompletionRequest,
     EmbeddingResponse,
 )
-from vllm.entrypoints.pooling.pooling.protocol import (
+from .pooling.protocol import (
     IOProcessorRequest,
     PoolingBytesResponse,
     PoolingChatRequest,
     PoolingCompletionRequest,
     PoolingResponse,
 )
-from vllm.entrypoints.pooling.scoring.protocol import ScoringRequest, ScoringResponse
-from vllm.entrypoints.pooling.scoring.typing import ScoringData
-from vllm.inputs import DataPrompt, EngineInput
-from vllm.lora.request import LoRARequest
+from .scoring.protocol import ScoringRequest, ScoringResponse
+from .scoring.typing import ScoringData
 
 PoolingCompletionLikeRequest: TypeAlias = (
     EmbeddingCompletionRequest
@@ -65,6 +66,8 @@ PoolingRequestT = TypeVar("PoolingRequestT", bound=AnyPoolingRequest)
 
 @dataclass(kw_only=True)
 class PoolingServeContext(Generic[PoolingRequestT]):
+    model_config = ConfigDict(arbitrary_types_allowed=True)
+
     request: PoolingRequestT
     raw_request: Request | None = None
     model_name: str
@@ -74,14 +77,14 @@ class PoolingServeContext(Generic[PoolingRequestT]):
     lora_request: LoRARequest | None = None
     engine_inputs: Sequence[EngineInput] | None = None
     prompt_request_ids: list[str] | None = None
-    intermediates: Any | None = None
 
     result_generator: AsyncGenerator[tuple[int, PoolingRequestOutput], None] | None = (
         None
     )
     final_res_batch: list[PoolingRequestOutput] = field(default_factory=list)
 
-    model_config = ConfigDict(arbitrary_types_allowed=True)
+    ## for Long Text Embedding with Chunked Processing
+    original_engine_inputs: Sequence[EngineInput] | None = None
 
     ## for bi-encoder & late-interaction
     n_queries: int | None = None
